+++ src/huffman/huffman/encoding/eof-encoding.cpp
@@ -4,10 +4,10 @@
 
 namespace {
 	class EofEncodingImplementation : public encoding::EncodingImplementation {
-		const u64 _domain_size;
+		const unsigned _domain_size;
 
 	public:
-		EofEncodingImplementation(const u64 domain_size) : _domain_size(domain_size) {}
+		EofEncodingImplementation(const unsigned domain_size) : _domain_size(domain_size) {}
 
 		void encode(io::InputStream& input_stream, io::OutputStream& output_stream) const override {
 			while (!input_stream.end_reached()) output_stream.write(input_stream.read());
@@ -15,7 +15,7 @@ namespace {
 		}
 
 		void decode(io::InputStream& input_stream, io::OutputStream& output_stream) const override {
-			const u64 eof = _domain_size;
+			const unsigned eof = _domain_size;
 
 			Datum datum = input_stream.read();
 
@@ -27,6 +27,6 @@ namespace {
 	};
 }
 
-std::shared_ptr<encoding::EncodingImplementation> encoding::create_eof_implementation(u64 domain_size) {
+std::shared_ptr<encoding::EncodingImplementation> encoding::create_eof_implementation(unsigned domain_size) {
 	return std::make_shared<EofEncodingImplementation>(domain_size);
 }
